+++ src/Event.hs
@@ -5,8 +5,8 @@ import           Data.Time
 import           Task
 
 data Event
-  = TaskCreated UTCTime Ref Id Pos Desc [Tag] Due
-  | TaskUpdated UTCTime Ref Id Pos Desc [Tag] Due
+  = TaskCreated UTCTime Ref Id Pos Desc [Tag] (Maybe Duration)
+  | TaskUpdated UTCTime Ref Id Pos Desc [Tag] (Maybe Duration)
   | TaskStarted UTCTime Ref Id
   | TaskStopped UTCTime Ref Id
   | TaskMarkedAsDone UTCTime Ref Id